[TASK] CI: Stop running SQL Server functional tests
Recent pdo_sqlsrv 5.10.0 force SSL enabled
connections with valid certificates. We can't
supply these in CI right now.

The patch removes active functional testing with
this DBMS on CI, but keeps the rest of the
infrastructure in runTests.sh and friends.

This is a quick fix to get v11 and v10 green
again. We could pick this up again if we find
a solution for the certificate issue.

Resolves: #96974
Releases: 11.5, 10.4
Change-Id: I0af99c52ae447ae6974d21236a862179e9b40dd6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73613
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>