[TASK] Replace most symfony/polyfill-* packages
The patch adds a couple of 'replace' entries for
common symfony/polyfill-* packages that are
obsolote due to our PHP version requirements
and PHP required extensions.

This is the recommended way as outlined
in symfony docs [1] at the moment.

Going this way in composer.json is semantically
wrong - TYPO3 does not in fact provide the
php functions in question itself. However, there
is currently no more elegant way to tell composer
that packages are obsolete automatically when
certain platform dependencies are met.

In the end, we don't really want to maintain
this list, but decided to still go with it with
v12 for now since the number of 'obsolete'
dependecies pile up and avoiding these packages
in projects can be a flat security advantage
(less code = less risk) and is environmental
more friendly.

[1] https://github.com/symfony/polyfill/#design

Resolves: #96980
Releases: main
Change-Id: I56e8c1da83c4a2c57aed1abbc30bde3aeb807355
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73623
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>