[TASK] Set "TSFE->requestedId" early in the process
The property "TSFE->requestedId" contains the
ID of the page that was initially requested.

Due to Site Handling and native URL Routing
since TYPO3 v9, this variable can be set initially
and not within determineId() / fetch_the_id(),
thus much earlier.

This way the property has no side-effects
when using $tsfe->getRequestedId() and just returns
the property as set in the constructor / setPageArguments().

Resolves: #96987
Releases: main
Change-Id: Ifd40cce4da9dd0139206d1306086a701608b889e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73634
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>