[BUGFIX] Use per-language runtime cache for LanguageService
LanguageService state has been moved into runtime cache
for further reducing the need to keep the LanguageService
in a global state. Failing to predict the need to translate
into multiple languages in the same context and process,
$cacheIdentifier had not be included the language.

This patch adds the current language to the $cacheIndentifier
to support multiple language translation while keeping the
LanguageService state clean.

Resolves: #97009
Related: #94415
Releases: main, 11.5
Change-Id: If1eaa3d2724ccdd12979a974c30d722878eb55d8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73680
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Jan Delius <me@jandelius.eu>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jan Delius <me@jandelius.eu>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>