[BUGFIX] Add missing comma to array_merge() in WorkspaceService
With deprecating 3rd party usage of getTreeList() in
QueryGenerator and QueryView with #92080 a comma has
been missed as element list separator for a list merge
with array_merge(...) in WorkspaceService. This emits
a corresponding native PHP warning or error.

Missing comma is added to avoid the warning and merge
the elements properly.

Resolves: #97011
Related: #92080
Releases: main, 11.5
Change-Id: I612eecac4ff1222191d6310e8996268e8746de69
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73682
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>