[TASK] Deprecate usage of RegularExpressionValidator
This patch deprecates the usage of the RegularExpressionValidator
within the form editor. The validator will be removed from the UI
in TYPO3 v13. Our goal is to unclutter and strip the form editor
from too technical and complex concepts. Regular expressions are
difficult to understand especially for our main target group which
are editors.

An according comment has been added to the configuration files of
the form framework.

Resolves: #97016
Releases: main
Change-Id: I895f2a377c545534869b79d93102544f2a84e36e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73689
Tested-by: Elias Häußler <e.haeussler@familie-redlich.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Elias Häußler <e.haeussler@familie-redlich.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>