[TASK] Deprecate order of validation message
The form framework ships a "date range validator". The validator can
for example be added via the form editor to the "date" form element.
Especially when it comes to this element, the order of the available
fields is not in line with all of the other form elements. The field
"Custom error message" (validationErrorMessage) is usually the last
field.

Since the order can only be adjusted by changing the key within the
form configuration, this patch adds according comments to the form
configuration. The breaking change will be done with TYPO3 13.0 to
further streamline the UI.

Resolves: #97019
Releases: main
Change-Id: Ia39e61a5a9abce22c1c327f0786c824851a06d63
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73691
Tested-by: core-ci <typo3@b13.com>
Tested-by: Elias Häußler <e.haeussler@familie-redlich.de>
Tested-by: Andreas Steiger <typo3@andreassteiger.de>
Tested-by: waldhacker <hello@waldhacker.dev>
Reviewed-by: Elias Häußler <e.haeussler@familie-redlich.de>
Reviewed-by: Andreas Steiger <typo3@andreassteiger.de>
Reviewed-by: waldhacker <hello@waldhacker.dev>