[TASK] styleguide: Add symmetric example with type group
https://github.com/TYPO3-CMS/styleguide/pull/171

A useful example with a symmetric field where type group is
used inside the mn table. styleguide already has an example
with type select, so now we have both options.
Type group e.g. is a good alternative to type select if you have
a huge amount of records.
There was an error with type group in a symmetric field which
has been fixed with https://forge.typo3.org/issues/92963