[TASK] Add request tests for DELETE,PATCH,PUT AND POST with data
The testing-framework with FE requests using sub requests
is now able to handle DELETE,PATCH,PUT and POST requests.

The situation in testing-framework is not the final solution,
though. The patch adds a test to verify if testing-framework
request details are properly received in the FE application.

This not only adds a use case to the core for these kind
of requests, but also ensures that we don't break this
detail again when testing-framework internal handling is
further refactored.

Resolves: #97084
Releases: main, 11.5
Change-Id: I8268625d4b439f1657168d6b9c9a3878b36477bd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73767
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>