[BUGFIX] Add record title of content in InternalLinkType
In case an internal link targets a content element,
defined by an anchor, this element is checked for
e.g. being moved, hidden or deleted.

In case the element is moved the "title" attribute
was not filled, leading to a PHP warning.

This is now fixed by always adding the correct
record title.

Resolves: #97147
Releases: main, 11.5
Change-Id: Ia78c34465bdc056a5714fc4e1bb7d7981be1e409
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73897
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>