[BUGFIX] Make full module information available
A couple of JS modules requires some module
information, such as the navigationComponent.

Those information were previously added as data
attributes to the main module menu items and then
retrieved by the JS components. This however did no longer
work for third-level modules, introduced in #97135,
since they are not present in the main module menu.

This is now resolved by adding the full module information,
including all modules, as a single data attribute to the
"scaffold-modulemenu" element.

Note: On installing new extensions, which add backend
modules, a complete backend refresh is necessary for
the new modules to be accessible. An automatic refresh will
be added in another patch, since this will also be
required to update the JavaScript import maps.

Resolves: #97184
Related: #97135
Releases: main
Change-Id: I753ef87f09ef111222945703ae3aa7cb2cb0d802
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73914
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>