[BUGFIX] Allow installation of incompatible distributions
TYPO3 allows to install incompatible extensions for a very long time,
but distributions were missed when this feature was introduced.
This patch now offers the same "Dependencies could not be resolved"
dialog for distributions that miss dependencies.

As this dialog is now used in different contexts, the "Go back" action
has been made configurable.

Resolves: #80219
Releases: main
Change-Id: If429164e539d9c3c984477cfdcb2568a59891117
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74024
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>