[TASK] Inform about not found user in record history
When a user is no longer available (e.g. deleted),
record history entries for this user were not clear
since only the general avatar without a name was
displayed. This was especially confusing in case
the "(via XYZ)" string was displayed, which happens
if the change was done in switch user mode.

This is improved by using the "[Not found]" label.

Additionally, it now also ensures the "(via XYZ)"
information is kept, in case the corresponding user
is no longer available.

Resolves: #97245
Releases: main, 11.5
Change-Id: I701098a07eab4969eefb24016f4351aef3de88bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74066
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>