[BUGFIX] Remove Symfony polyfill replacements from core
Adding a replace to a package semantically means, that
this package replaces another one. This is not true
for typo3/cms-core and therefore the replacement should
not be declared.

Additionally, composer fails when more than one dependency
declare identical replaces. We thus can not declare them
in typo3/cms-core since it would conflict as soon as any
other dependency does that, too.

It is fine for typo3/cms as this only serves as base for
a bundled build for non Composer installations and for
development. As such, the replacements are kept in
monorepo root composer.json.

Resolves: #97295
Related: #96980
Releases: main
Change-Id: I7a65182975457d71f611bf102c2842c6972e4eba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74169
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>