[BUGFIX] Handle flexform in BU::getProcessedValue() only if not null
The contents of a field of type `flex` may be NULL if the field is
empty. `BackendUtility::getProcessedValue()` handles the content of such
`flex` fields now only if the value is an actual string to solve
deprecations related to PHP 8.1.

In the same run, the content is trimmed to get rid of lots of empty
lines caused by multiline XML data.

Resolves: #97297
Releases: main
Change-Id: I4261d16bfa819fd1151bbe6905b97f2571701fee
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74174
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>