[FEATURE] Allow keyboard navigation in live search
Currently it's not possible to select a suggestion
of the live search.

This has been changed to add a highlight style and
load the content on "enter/tab"-key pressed.

This also tries to simplify the entire structure
which used to be a combination of bootstraps dropdown
and autocomplete which had to be in sync to show
and hide which turned out to be simply a mess.

Resolves: #97347
Releases: main
Change-Id: I74895c98560bd8cbe32ce5caca2974893f320710
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74257
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>