[BUGFIX] Translate additional scheduler field labels in controller
The "langLabel" field might be already translated
instead of a "LLL-string". This causes the
TranslateViewHelper to throw an exception, because
it doesn't know what to do.

The better way is to translate this field already
in the controller, where LanguageService can be used
directly, without failing if no extension key can
be found.

Resolves: #97349
Related: #97312
Releases: main
Change-Id: I06f759b1c0961a26e5236d576948aecc1acc0d9b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74261
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>