[BUGFIX] Fix sorting of file properties in FileList
Since the introduction of the column selector to the
FileList (#96959), it had been possible to sort by
all those selected columns. However, this actually
never worked, because the sorting was done in the
driver (e.g. LocalDriver), which does not know
anything about the file (metadata) properties.
This is also the reason why sorting of folders
always just lead them to jump around.

To fix this, sorting is now done in the FileList,
where working on the file/folder objects directly
is possible. This also now finally ensures that
the displayed value is the same value which is
used for sorting. This previously was not always
the case, see e.g. #97261 or #97290.

Additionally, this patch fixes some further sorting
related bugs, e.g. the control column header is no
longer wrapped into a sort link.

Resolves: #97443
Related: #96959
Releases: main, 11.5
Change-Id: I24733d8137e25a9f64c3262a9624bdcb986441e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74372
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>