[TASK] Migrate form-horizontal class to Bootstrap 5
This patch migrates the "form-horizontal" class from
Bootstrap 3 to the new style by using the "row" and
"col-form-label" classes.

Usages of form-group have also been replaced by the
utility "mb-3" class. This allowed to drop one Css
rule for removing the margin of the last element in
the LinkBrowser.

Places adjusted:
* Settings > Presets
* Form > Duplicate
* LinkBrowser
* ExtensionManager > Upload extension

Some places have already been migrated, but forgot
to remove the class.

In addition, the id "typo3-linkURL" has been removed
along the way, as the styling was removed in #71840
already.

Resolves: #97405
Related: #92616
Related: #71840
Releases: main, 11.5
Change-Id: I20b9e83acf9cca46d7cb8a0a3826ae526b3e2b3a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74335
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>