[FEATURE] Introduce configurable password policies
TYPO3 now includes a PasswordPolicyValidator component which can
be used to validate passwords against configurable password policies.
TYPO3 also includes a default password policy which ensures that
passwords meet strong security defaults.

Password policies can be configured individually for both frontend
and backend context. It is also possible to extend a password policy
with own validation requirements.

As a first step, the included default password policy is applied to
ext:setup to ensure, that new passwords of backend users entered in
"User Settings" will match the default password requirements.

This patch is the first of a series of patches to implement
password policies globally in TYPO3. See issue #97387 for details.

Resolves: #97388
Releases: main
Change-Id: I6cb6ca77e7c19b719fe36c0c51967e31bc41082e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74350
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>