[BUGFIX] Name invalid identifier in UpgradeWizard exception
An exception will be raised, if an upgrade wizard identifier
returned by "UpgradeWizardInterface::getIdentifier()" is not
valid due to e.g. misspelling.

The invalid identifier is now mentioned in the exception
message for easier debugging.

In addition, the exception is now split up into two separate
exceptions in order to define the reason more precisely.

Resolves: #97437
Releases: main, 11.5
Change-Id: Ie34ad4e4bb698109e224488ff1b43e2b3d66e7fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74392
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>