[BUGFIX] Replace array notation with string for value of property allowedAttribs
According to \TYPO3\CMS\Core\Html\HtmlParser::HTMLcleaner() the value of property "allowedAttribs" needs to be a commalist of attributes (string value) instead of an array. In addition, tests have shown that the class attribute for hr tags is currently removed despite the specification. This bugfix ensures that the class attribute is not stripped by the parser as originally intended.